Fix validation on search input
The regex was updated on the search bar input to allow for a user to
enter either a city or a city and state (US only) or city and country
code.

A return value was given to the validation function in the JS file. This
return value is used to halt execution unless a valid input is entered.